requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
Django==5.2.4
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-htmx==1.23.2
djangorestframework==3.16.0
griffe==1.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
Markdown==3.8.2
mcp==1.12.3
nest-asyncio==1.6.0
numpy==2.3.2
openai==1.99.1
openai-agents==0.2.4
packaging==25.0
pdf2image==1.17.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg-binary==3.2.9
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
redis==6.2.0
referencing==0.36.2
reportlab==4.4.3
requests==2.32.4
rpds-py==0.26.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.47.2
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
